requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.5
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
